+++ pkg/rpc/client/neoscan.go
@@ -1,4 +1,4 @@
-package rpc
+package client
 
 import (
 	"encoding/json"
